requirements.txt
# Python FastAPI Chatbot Service Dependencies

fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Google Gemini
google-generativeai==0.8.3

# Vector Database Options
# Option 1: Pinecone (Cloud-based)
pinecone-client==2.2.4

# Option 2: FAISS (Local) - Updated to available version
faiss-cpu==1.9.0.post1
# For GPU: faiss-gpu==1.7.4

# Embeddings
sentence-transformers==2.2.2

# Database - Using psycopg2 for Python 3.13 compatibility
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Utilities
numpy>=1.24.3
pandas>=2.1.4
python-multipart==0.0.6
pydantic-settings==2.1.0

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
